+++ src/main-personal-settings.ts
@@ -1,9 +1,9 @@
 import Vue from 'vue';
 import { createPinia, PiniaVuePlugin } from 'pinia';
 
-import l10n from '@nextcloud/l10n';
+import './l10n';
 
-import './main.css';
+import './css/main.css';
 
 import '@nextcloud/dialogs/styles/toast.scss';
 
@@ -13,9 +13,6 @@ import PersonalSettings from './components/PersonalSettings.vue';
 
 Vue.use(PiniaVuePlugin);
 const pinia = createPinia();
-
-Vue.prototype.t = l10n.translate;
-Vue.prototype.n = l10n.translatePlural;
 
 new Vue({
   el: '#settings-personal-money',